Minor improvements

+++ classes/CoveredAreaCondition.php
@@ -15,8 +15,8 @@ class CoveredAreaCondition
     public function __construct(array $condition = [])
     {
         $this->type = array_get($condition, 'type', 'all');
-        $this->amount = (int)array_get($condition, 'amount', -1);
-        $this->total = (int)array_get($condition, 'total', 0);
+        $this->amount = (float)array_get($condition, 'amount', -1);
+        $this->total = (float)array_get($condition, 'total', 0);
         $this->priority = (int)array_get($condition, 'priority', 999);
     }
